scripts/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
qdrant-client==1.7.3
numpy==1.26.4
pydantic==2.6.1
redis==5.0.1
python-dotenv==1.0.1
flask==3.0.3
feedparser==6.0.11
sentence-transformers==2.7.0
newspaper3k==0.2.8
lxml_html_clean
